split URI classification helpers into dedicated microcrate
