+++ crates/perl-uri-classify/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "perl-uri-classify"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "URI classification and key normalization helpers for Perl LSP"
+readme = "README.md"
+documentation = "https://docs.rs/perl-uri-classify"
+keywords = ["perl", "uri", "lsp", "classification", "normalization"]
+categories = ["development-tools", "encoding"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[dependencies]
+url = "2.5.8"
+
+[lints]
+workspace = true
