+++ crates/perl-uri/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["development-tools", "encoding"]
 include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
 
 [dependencies]
+perl-uri-classify = { workspace = true }
 url = "2.5.8"
 
 [dev-dependencies]
