+++ crates/perl-uri/src/lib.rs
@@ -198,120 +198,7 @@ pub fn normalize_uri(uri: &str) -> String {
     if let Ok(url) = Url::parse(uri) { url.to_string() } else { uri.to_string() }
 }
 
-/// Normalize a URI to a consistent key for lookups.
-///
-/// This function handles platform-specific differences to ensure consistent
-/// lookups across different systems, particularly for Windows drive letters.
-///
-/// # Windows Drive Letter Normalization
-///
-/// On Windows, drive letters in URIs may be uppercase or lowercase.
-/// This function normalizes them to lowercase for consistent lookups:
-/// - `file:///C:/foo` → `file:///c:/foo`
-/// - `file:///D:/bar` → `file:///d:/bar`
-///
-/// # Examples
-///
-/// ```
-/// use perl_uri::uri_key;
-///
-/// // Standard URI
-/// let key = uri_key("file:///tmp/test.pl");
-/// assert_eq!(key, "file:///tmp/test.pl");
-///
-/// // Windows URI with uppercase drive
-/// let key = uri_key("file:///C:/Users/test.pl");
-/// assert_eq!(key, "file:///c:/Users/test.pl");
-///
-/// // Invalid URI returned as-is
-/// let key = uri_key("not-a-uri");
-/// assert_eq!(key, "not-a-uri");
-/// ```
-pub fn uri_key(uri: &str) -> String {
-    if let Ok(u) = Url::parse(uri) {
-        let s = u.as_str().to_string();
-        if let Some(rest) = s.strip_prefix("file:///") {
-            // Check for Windows drive letter pattern: single letter followed by colon
-            if rest.len() > 1
-                && rest.as_bytes()[1] == b':'
-                && rest.as_bytes()[0].is_ascii_alphabetic()
-            {
-                // Normalize drive letter to lowercase
-                return format!("file:///{}{}", rest[0..1].to_ascii_lowercase(), &rest[1..]);
-            }
-        }
-        s
-    } else {
-        uri.to_string()
-    }
-}
-
-/// Check if a URI uses the `file://` scheme.
-///
-/// # Examples
-///
-/// ```
-/// use perl_uri::is_file_uri;
-///
-/// assert!(is_file_uri("file:///tmp/test.pl"));
-/// assert!(!is_file_uri("https://example.com"));
-/// assert!(!is_file_uri("untitled:Untitled-1"));
-/// ```
-pub fn is_file_uri(uri: &str) -> bool {
-    uri.starts_with("file://")
-}
-
-/// Check if a URI uses a special scheme (not `file://`).
-///
-/// Special schemes include:
-/// - `untitled:` - Unsaved documents
-/// - `vscode-notebook:` - VS Code notebooks
-/// - `git:` - Git diff views
-/// - etc.
-///
-/// # Examples
-///
-/// ```
-/// use perl_uri::is_special_scheme;
-///
-/// assert!(is_special_scheme("untitled:Untitled-1"));
-/// assert!(is_special_scheme("git:/foo/bar"));
-/// assert!(!is_special_scheme("file:///tmp/test.pl"));
-/// ```
-pub fn is_special_scheme(uri: &str) -> bool {
-    if let Ok(url) = Url::parse(uri) {
-        url.scheme() != "file"
-    } else {
-        // If it can't be parsed as a URL, check for common special prefixes
-        uri.starts_with("untitled:")
-            || uri.starts_with("git:")
-            || uri.starts_with("vscode-notebook:")
-            || uri.starts_with("vscode-vfs:")
-    }
-}
-
-/// Extract the file extension from a URI.
-///
-/// # Examples
-///
-/// ```
-/// use perl_uri::uri_extension;
-///
-/// assert_eq!(uri_extension("file:///tmp/test.pl"), Some("pl"));
-/// assert_eq!(uri_extension("file:///tmp/Module.pm"), Some("pm"));
-/// assert_eq!(uri_extension("file:///tmp/no-extension"), None);
-/// ```
-pub fn uri_extension(uri: &str) -> Option<&str> {
-    // Find the last path segment
-    let path_part = uri.rsplit('/').next()?;
-    // Remove query string and fragment
-    let path_part = path_part.split('?').next()?;
-    let path_part = path_part.split('#').next()?;
-    // Find the extension
-    let dot_pos = path_part.rfind('.')?;
-    let ext = &path_part[dot_pos + 1..];
-    if ext.is_empty() { None } else { Some(ext) }
-}
+pub use perl_uri_classify::{is_file_uri, is_special_scheme, uri_extension, uri_key};
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used, clippy::expect_used)]
